[REFACTOR] cache translations, yaml is slow
With every single unit test that translated error messages this function
was hit, which would parse the yaml files from disk. A simple cache
makes this 10x as fast, shaving about 15 seconds from the CI run.

When we work on the language locally and introduce new keywords, we
might now need to restart the process, but I believe this should
already be covered by flask.